Windows Fixes
Oops... Looks like I broke POSTAL on Windows with all those 64-bit
changes. I think this is fixed now, but I don't have a working
Visual Studio environment on my Windows VM. Please test.

+++ RSPiX/Src/GREEN/Hot/hot.cpp
@@ -66,7 +66,7 @@
 //							dequeued or polled events, but it was getting stale any-
 //							ways.
 //
-//		01/26/97	JMI	Changed callbacks' first parms to RHot* instead of ULONG.
+//		01/26/97	JMI	Changed callbacks' first parms to RHot* instead of uint32_t.
 //
 //		03/19/97	JMI	Added InputEventCall and made Do(pie) the main RHot
 //							interface.
@@ -153,7 +153,7 @@ RHot::RHot(
 	int16_t sH,										// Height of new hotbox.
 	REventCall fnEventCall /*= NULL*/,		// Callback on mouse event.
 	int16_t	sActive /*= FALSE*/,					// Initially active, if TRUE.
-	ULONG	ulUser /*= 0*/,						// User value.
+	uint32_t	ulUser /*= 0*/,						// User value.
 	int16_t sPriority /*= RHOT_NO_PRIORITY*/)// Priority.  Default == non-prioritized.
 	{
 	// Reset all members.
@@ -186,7 +186,7 @@ RHot::RHot(
 	int16_t sH,										// Height of new hotbox.
 	REventPosCall fnEventPosCall,				// Callback on mouse event.
 	int16_t	sActive /*= FALSE*/,					// Initially active, if TRUE.
-	ULONG	ulUser /*= 0*/,						// User value.
+	uint32_t	ulUser /*= 0*/,						// User value.
 	int16_t sPriority /*= RHOT_NO_PRIORITY*/)// Priority.  Default == non-prioritized.
 	{
 	// Reset all members.
@@ -219,7 +219,7 @@ RHot::RHot(
 	int16_t sH,										// Height of new hotbox.
 	InputEventCall fnInputEventCall,			// Callback on mouse event.
 	int16_t	sActive /*= FALSE*/,					// Initially active, if TRUE.
-	ULONG	ulUser /*= 0*/,						// User value.
+	uint32_t	ulUser /*= 0*/,						// User value.
 	int16_t sPriority /*= RHOT_NO_PRIORITY*/)// Priority.  Default == non-prioritized.
 	{
 	// Reset all members.
